+++ src/main/java/org/sopt/makers/internal/report/service/SoptReportStatsService.java
@@ -13,7 +13,7 @@ import java.util.stream.Collectors;
 
 import org.sopt.makers.internal.community.repository.CommunityPostLikeRepository;
 import org.sopt.makers.internal.community.repository.CommunityPostRepository;
-import org.sopt.makers.internal.domain.Word;
+import org.sopt.makers.internal.wordchaingame.domain.Word;
 import org.sopt.makers.internal.external.MakersCrewClient;
 import org.sopt.makers.internal.report.domain.PlaygroundType;
 import org.sopt.makers.internal.report.domain.SoptReportCategory;
@@ -23,8 +23,8 @@ import org.sopt.makers.internal.report.dto.PlayGroundTypeStatsDto;
 import org.sopt.makers.internal.report.dto.response.MySoptReportStatsResponse;
 import org.sopt.makers.internal.report.repository.AmplitudeEventRawDataRepository;
 import org.sopt.makers.internal.report.repository.SoptReportStatsRepository;
-import org.sopt.makers.internal.repository.WordChainGameQueryRepository;
-import org.sopt.makers.internal.repository.WordRepository;
+import org.sopt.makers.internal.wordchaingame.repository.WordChainGameQueryRepository;
+import org.sopt.makers.internal.wordchaingame.repository.WordRepository;
 import org.sopt.makers.internal.repository.community.CommunityCommentRepository;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
