+++ src/main/java/org/sopt/makers/internal/wordchaingame/dto/response/WordChainGameAllResponse.java
@@ -1,4 +1,4 @@
-package org.sopt.makers.internal.dto.wordChainGame;
+package org.sopt.makers.internal.wordchaingame.dto.response;
 
 import java.util.List;
