+++ src/main/java/org/sopt/makers/internal/wordchaingame/dto/response/WordChainGameRoomLastwordResponse.java
@@ -1,6 +1,4 @@
-package org.sopt.makers.internal.dto.wordChainGame;
-
-import java.util.List;
+package org.sopt.makers.internal.wordchaingame.dto.response;
 
 public record WordChainGameRoomLastwordResponse(
         Long roomId,
