+++ src/main/java/org/sopt/makers/internal/wordchaingame/repository/WordChainGameQueryRepository.java
@@ -1,12 +1,15 @@
-package org.sopt.makers.internal.repository;
+package org.sopt.makers.internal.wordchaingame.repository;
 
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import lombok.val;
-import org.sopt.makers.internal.domain.*;
-import org.sopt.makers.internal.dto.wordChainGame.QWinnerDao;
-import org.sopt.makers.internal.dto.wordChainGame.WinnerDao;
+import org.sopt.makers.internal.wordchaingame.domain.QWordChainGameRoom;
+import org.sopt.makers.internal.wordchaingame.domain.QWordChainGameWinner;
+import org.sopt.makers.internal.wordchaingame.domain.WordChainGameRoom;
+import org.sopt.makers.internal.wordchaingame.dto.response.QWinnerDao;
+import org.sopt.makers.internal.wordchaingame.dto.response.WinnerDao;
+import org.sopt.makers.internal.domain.QMember;
 import org.springframework.stereotype.Repository;
 
 import java.time.LocalDateTime;
