+++ src/main/java/org/sopt/makers/internal/wordchaingame/repository/WordChainGameRepository.java
@@ -1,6 +1,6 @@
-package org.sopt.makers.internal.repository;
+package org.sopt.makers.internal.wordchaingame.repository;
 
-import org.sopt.makers.internal.domain.WordChainGameRoom;
+import org.sopt.makers.internal.wordchaingame.domain.WordChainGameRoom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface WordChainGameRepository extends JpaRepository<WordChainGameRoom, Long> {
