+++ src/main/java/org/sopt/makers/internal/wordchaingame/repository/WordChainGameWinnerRepository.java
@@ -1,6 +1,6 @@
-package org.sopt.makers.internal.repository;
+package org.sopt.makers.internal.wordchaingame.repository;
 
-import org.sopt.makers.internal.domain.WordChainGameWinner;
+import org.sopt.makers.internal.wordchaingame.domain.WordChainGameWinner;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface WordChainGameWinnerRepository extends JpaRepository<WordChainGameWinner, Long> {
